Code cleanups and improvements.
