+++ NitroxReloader/ReloadableMethod.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace NitroxReloader
+{
+    [AttributeUsage(AttributeTargets.Method)]
+    public class ReloadableMethodAttribute : Attribute
+    {
+    }
+}
